Introduce GaeApi class and refactor code
