+++ src/main/java/teammates/common/util/GaeApi.java
@@ -1,0 +1,36 @@
+package teammates.common.util;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+import com.google.appengine.api.modules.ModulesService;
+import com.google.appengine.api.modules.ModulesServiceFactory;
+
+/**
+ * Calls API functions of Google AppEngine
+ */
+public class GaeApi {
+    private static ModulesService modulesService = ModulesServiceFactory.getModulesService();
+    
+    /**
+     * Gets all available versions.
+     */
+    public static List<Version> getAvailableVersions() {
+        List<String> versionListInString = new ArrayList<String>(modulesService.getVersions(null)); // null == default module
+        List<Version> versionList = new ArrayList<Version>();
+        for(String versionInString : versionListInString) {
+            versionList.add(new Version(versionInString));
+        }
+        Collections.sort(versionList);
+        return versionList;
+    }
+    
+    /**
+     * Gets the current version of the application.
+     */
+    public static Version getCurrentVersion() {
+        return new Version(modulesService.getCurrentVersion());
+    }
+    
+}
