+++ src/main/java/teammates/common/util/Version.java
@@ -1,18 +1,10 @@
 package teammates.common.util;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-
-import com.google.appengine.api.modules.ModulesService;
-import com.google.appengine.api.modules.ModulesServiceFactory;
-
 /**
  * Represents a version by 3 parts: major version, minor version and patch version.
  */
-public class Version {
-    private static ModulesService modulesService = ModulesServiceFactory.getModulesService();
+public class Version implements Comparable<Version> {
+    private String originalRepresentation;
     private Integer major;
     private Integer minor;
     private String patch;
@@ -45,6 +37,7 @@ public class Version {
      * It accepts either XX-XX-XXXXX or XX.XX.XXXX format.
      */
     public Version(String versionInString) {
+        originalRepresentation = versionInString;
         String[] list;
         if (versionInString.contains("-")) {
             list = versionInString.split("-", 3);   // split into at most 3 parts
@@ -80,114 +73,78 @@ public class Version {
      * Converts Version to String in format XX.XX.XXXX
      */
     public String toString() {
-        String version = "";
-        if (major != null) {
-            version += String.valueOf(major);
-        }
-        if (minor != null) {
-            String minorInString = String.valueOf(minor);
-            if (minorInString.length() < 2) {
-                minorInString = "0" + minorInString;
-            }
-            version += "." + minorInString;
-        }
-        if (patch != null) {
-            version += "." + patch;
-        }
-        return version;
+        return originalRepresentation.replaceAll("-", ".");
     }
     
     /**
      * Converts to String in format XX-XX-XXXX
      */
     public String toStringForQuery() {
-        return toString().replace('.', '-');
+        return originalRepresentation;
+    }
+    
+    private int compareVersionNumber(Integer num1, Integer num2) {
+        if (num1 == null && num2 == null) {
+            return 0;
+        }
+        if (num1 == null) {
+            return 1;
+        }
+        if (num2 == null) {
+            return -1;
+        }
+        return -num1.compareTo(num2);
+    }
+    
+    private int compareVersionString(String s1, String s2) {
+        if (s1 == null && s2 == null) {
+            return 0;
+        }
+        if (s1 == null) {
+            return 1;
+        }
+        if (s2 == null) {
+            return -1;
+        }
+        return -s1.compareTo(s2);
     }
     
     /**
-     * Gets all available versions.
+     * Compares versions by major, minor then by patch.
+     * The version with greater major, minor or patch will be smaller.
      */
-    public static List<Version> getAvailableVersions() {
-        List<String> versionListInString = new ArrayList<String>(modulesService.getVersions(null)); // null == default module
-        List<Version> versionList = new ArrayList<Version>();
-        for(String versionInString : versionListInString) {
-            versionList.add(new Version(versionInString));
+    @Override
+    public int compareTo(Version anotherVersion) {
+        int majorComparisonResult = compareVersionNumber(this.getMajorVersion(), anotherVersion.getMajorVersion());
+        if (majorComparisonResult != 0) {
+            return majorComparisonResult;
         }
-        Collections.sort(versionList, new VersionComparator());
-        return versionList;
-    }
-    
-    /**
-     * Comparator for version strings.
-     * It sorts versions in DESCENDING order by major version then by minor version then by patch version
-     */
-    private static class VersionComparator implements Comparator<Version> {
-        private int compareVersionNumber(Integer num1, Integer num2) {
-            if (num1 == null && num2 == null) {
-                return 0;
-            }
-            if (num1 == null) {
-                return 1;
-            }
-            if (num2 == null) {
-                return -1;
-            }
-            return -num1.compareTo(num2);
+        int minorComparisonResult = compareVersionNumber(this.getMinorVersion(), anotherVersion.getMinorVersion());
+        if (minorComparisonResult != 0) {
+            return minorComparisonResult;
         }
-        
-        private int compareVersionString(String s1, String s2) {
-            if (s1 == null && s2 == null) {
-                return 0;
-            }
-            if (s1 == null) {
-                return 1;
-            }
-            if (s2 == null) {
-                return -1;
-            }
-            return -s1.compareTo(s2);
-        }
-        
-        @Override
-        public int compare(Version v1, Version v2) {
-            int majorComparisonResult = compareVersionNumber(v1.getMajorVersion(), v2.getMajorVersion());
-            if (majorComparisonResult != 0) {
-                return majorComparisonResult;
-            }
-            int minorComparisonResult = compareVersionNumber(v1.getMinorVersion(), v2.getMinorVersion());
-            if (minorComparisonResult != 0) {
-                return minorComparisonResult;
-            }
-            int patchComparisonResult = compareVersionString(v1.getPatchVersion(), v2.getPatchVersion());
-            return patchComparisonResult;
-        }
-    }
-    
-    /**
-     * Gets the current version of the application.
-     */
-    public static Version getCurrentVersion() {
-        return new Version(modulesService.getCurrentVersion());
+        int patchComparisonResult = compareVersionString(this.getPatchVersion(), anotherVersion.getPatchVersion());
+        return patchComparisonResult;
     }
     
     /**
      * Gets the major version number of the version.
      */
-    private Integer getMajorVersion() {
+    public Integer getMajorVersion() {
         return major;
     }
     
     /**
      * Gets the minor version number of the version.
      */
-    private Integer getMinorVersion() {
+    public Integer getMinorVersion() {
         return minor;
     }
     
     /**
      * Gets the patch part of the version.
      */
-    private String getPatchVersion() {
+    public String getPatchVersion() {
         return patch;
     }
 }
